Apply hydroUplandWeighting to all worldclim variables

+++ runUplandWeighting_batch.py
@@ -13,16 +13,9 @@ weightras = os.path.join(geomdir,'Accu_area_grids', 'pixel_area_skm_15s.gdb', 'p
 uplandras = os.path.join(geomdir, 'Accu_area_grids', 'upstream_area_skm_15s.gdb', 'up_area_skm_15s')
 
 #Value grids
-worldclim_value = getfilelist(os.path.join(resdir, 'worldclimv2.gdb'), '.*nibble.*')
+worldclim_value = getfilelist(os.path.join(resdir, 'worldclimv2.gdb'), '.*nibble$')
 
-value_grid = worldclim_value[0]
-direction_grid = directionras
-weight_grid = weightras
-upland_grid = uplandras
-scratch_dir = os.path.join(resdir, 'scratch.gdb')
-out_dir = os.path.join(resdir, 'worldclimv2.gdb')
-
-def hydroUplandWeighting(value_grid, direction_grid, weight_grid, upland_grid, scratch_dir, out_dir):
+def hydroUplandWeighting(value_grid, direction_grid, weight_grid, upland_grid, scratch_dir, out_dir, overwrite=False):
     #Check that directories exist, otherwise, create them
     pathcheckcreate(scratch_dir)
     pathcheckcreate(out_dir)
@@ -35,38 +28,47 @@ def hydroUplandWeighting(value_grid, direction_grid, weight_grid, upland_grid, s
     xpxarea_ac1 = os.path.join(out_dir, '{}_xpxarea_ac1'.format(nameroot))
     xpxarea_ac_fin = os.path.join(out_dir, '{}_xpxarea_ac_fin'.format(nameroot))
 
-    try:
-        set_environment(out_dir, direction_grid)
+    if (not arcpy.Exists(out_grid)) or (overwrite == True):
+        try:
+            set_environment(out_dir, direction_grid)
 
-        # Multiply input grid by pixel area
-        print('Processing {}...'.format(xpxarea))
-        valueXarea = Times(r"" + value_grid, r"" + weight_grid)
-        valueXarea.save(xpxarea)
+            # Multiply input grid by pixel area
+            print('Processing {}...'.format(xpxarea))
+            valueXarea = Times(r"" + value_grid, r"" + weight_grid)
+            valueXarea.save(xpxarea)
 
-        # Flow accumulation of value grid and pixel area product
-        print('Processing {}...'.format(xpxarea_ac1))
-        outFlowAccumulation = FlowAccumulation(direction_grid, xpxarea, "FLOAT")
-        outFlowAccumulation.save(xpxarea_ac1)
+            # Flow accumulation of value grid and pixel area product
+            print('Processing {}...'.format(xpxarea_ac1))
+            outFlowAccumulation = FlowAccumulation(direction_grid, xpxarea, "FLOAT")
+            outFlowAccumulation.save(xpxarea_ac1)
 
-        print('Processing {}...'.format(xpxarea_ac_fin))
-        outFlowAccumulation_2 = Plus(xpxarea_ac1, xpxarea)
-        outFlowAccumulation_2.save(xpxarea_ac_fin)
+            print('Processing {}...'.format(xpxarea_ac_fin))
+            outFlowAccumulation_2 = Plus(xpxarea_ac1, xpxarea)
+            outFlowAccumulation_2.save(xpxarea_ac_fin)
 
-        # Divide by the accumulated pixel area grid
-        print('Processing {}...'.format(out_grid))
-        UplandGrid = Divide(xpxarea_ac_fin, upland_grid)
-        UplandGrid.save(out_grid)
+            # Divide by the accumulated pixel area grid
+            print('Processing {}...'.format(out_grid))
+            UplandGrid = Divide(xpxarea_ac_fin, upland_grid)
+            UplandGrid.save(out_grid)
 
-    except Exception, e:
-        # If an error occurred, print line number and error message
-        import traceback, sys
-        tb = sys.exc_info()[2]
-        arcpy.AddMessage("Line %i" % tb.tb_lineno)
-        arcpy.AddMessage(str(e.message))
+        except Exception, e:
+            # If an error occurred, print line number and error message
+            import traceback, sys
+            tb = sys.exc_info()[2]
+            arcpy.AddMessage("Line %i" % tb.tb_lineno)
+            arcpy.AddMessage(str(e.message))
 
-    print('Deleting intermediate grids...')
-    for lyr in [xpxarea, xpxarea_ac1, xpxarea_ac_fin]:
-        if arcpy.Exists(lyr):
-            arcpy.Delete_management(lyr)
+        print('Deleting intermediate grids...')
+        for lyr in [xpxarea, xpxarea_ac1, xpxarea_ac_fin]:
+            if arcpy.Exists(lyr):
+                arcpy.Delete_management(lyr)
+    else:
+        print('{} already exists and overwrite==False...'.format(out_grid))
 
-hydroUplandWeighting(value_grid, direction_grid, weight_grid, upland_grid, scratch_dir, out_dir)
+for wc_valuegrid in worldclim_value:
+    hydroUplandWeighting(value_grid = wc_valuegrid,
+                         direction_grid = directionras,
+                         weight_grid = weightras,
+                         upland_grid = uplandras,
+                         scratch_dir = os.path.join(resdir, 'scratch.gdb'),
+                         out_dir = os.path.join(resdir, 'worldclimv2.gdb'))
